Extract result_set pagination, sorting and filtering to ReportResultSet mixin
